Refactor news articles
