+++ backend/routes/api/newsArticle/update.js
@@ -2,16 +2,17 @@ const express = require('express');
 
 const router = express.Router();
 const { body, validationResult } = require('express-validator');
-const { authenticateToken } = require('../../../middlewares/auth');
-const sanitize = require('sanitize-filename')
+const sanitize = require('sanitize-filename');
 const upload = require('multer')({ dest: 'uploads/', limits: { fieldSize: 50 * 1024 * 1024 } });
-const { updateNewsArticleById } = require('../../../repository/newsArticle');
 const mongoose = require('mongoose');
+
 const Image = mongoose.model('Image');
-const { updateImageById } = require('../../../repository/image');
 const fs = require('fs');
 const path = require('path');
 const escape = require('escape-html');
+const { updateImageById } = require('../../../repository/image');
+const { updateNewsArticleById } = require('../../../repository/newsArticle');
+const { authenticateToken } = require('../../../middlewares/auth');
 
 router.use(express.json());
 
@@ -24,7 +25,9 @@ router.post('/', [
       return res.status(400).json({ errors: errors.array() });
     }
 
-    const { _id: id, title, originalImageId, content } = req.body;
+    const {
+      _id: id, title, originalImageId, content,
+    } = req.body;
     if (!id || !title || !content) {
       return res.status(500).send('Could not create news article.');
     }
